Add PEP-517 compliant build requirement specification
PEP 517 [1] provides a new standard format for specifying build-time
dependencies for Python packages. This format supercedes previous
formats, such as the `setup_requirements` argument in setuptools[2].

In order for this package to be installable by systems that expect PEP
517-compliant formatting, such as Poetry, a minimal pyproject.toml file
must be added specifying the build-time requirements. This file does not
trigger any other changes to the build system or require the adoption of
any new tools; recent versions of setuptools support referencing the
pyproject.toml `build-system` header information.

[1]: https://peps.python.org/pep-0517/
[2]: https://setuptools.pypa.io/en/latest/userguide/dependency_management.html#build-system-requirement